+++ include/coap2/net.h
@@ -739,4 +739,15 @@ coap_pdu_t *coap_wellknown_response(coap_context_t *context,
  */
 unsigned int coap_calc_timeout(coap_session_t *session, unsigned char r);
 
+/**
+ * Function interface for joining a multicast group for listening
+ *
+ * @param ctx   The current context
+ * @param groupname The name of the group that is to be joined for listening
+ *
+ * @return       0 on success, -1 on error
+ */
+int
+coap_join_mcast_group(coap_context_t *ctx, const char *groupname);
+
 #endif /* COAP_NET_H_ */
