+++ man/coap-rd.txt.in
@@ -61,10 +61,10 @@ verbosity level is set to '5'.
 
 * Example
 ----
-coap-rd -A 2001:db8:81a8:0:6ef0:dead:feed:beef  -g FF02:FD
+coap-rd -A 2001:db8:81a8:0:6ef0:dead:feed:beef  -g FF02::FD
 ----
 Set listening address to '2001:db8:81a8:0:6ef0:dead:feed:beef' and join the
-All CoAP Nodes multicast group 'FF02:FD'.
+All CoAP Nodes multicast group 'FF02::FD'.
 
 FILES
 ------
